Refactor JavaScript build process with TypeScript
